requirements.txt
SQLAlchemy==1.2.13
Flask_Moment==0.6.0
Flask_SQLAlchemy==2.3.2
Flask_Login==0.4.1
Werkzeug==0.15.3
alembic==1.0.2
Flask_Bcrypt==0.7.1
Flask_WTF==0.14.2
WTForms==2.2.1
Flask==1.0.2
Flask_Mail==0.9.1
Flask_Migrate==2.3.0
python-dotenv==0.10.0
PyJWT==1.7.0
Flask-Markdown==0.3
gunicorn==19.9.0
psycopg2==2.7.6.1
bleach
requests
nltk